requirements.txt
fastapi==0.115.12
uvicorn==0.30.1
sqlalchemy==2.0.27
pydantic==2.11.0
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
email-validator==2.1.1
alembic==1.13.1